requirements.txt
PyQt5==5.15.4
pywinauto==0.6.8
opencv-python==4.5.3.56
numpy==1.21.2
pyautogui==0.9.53
Pillow==8.3.1
